Use colon recipe prefix with guard in tests
