+++ src/glide_session.h
@@ -2,39 +2,34 @@
 
 #include "glide_process.h"
 #include "interaction.h"
+#include "status_service.h"
 
 #include <glib.h>
 
 typedef struct _GlideSession GlideSession;
 typedef void (*GlideSessionCallback)(GlideSession *self, Interaction *interaction, gpointer user_data);
 
-typedef struct {
-  void (*eval)(GlideSession *self, Interaction *interaction);
-  void (*set_interaction_added_cb)(GlideSession *self, GlideSessionCallback cb, gpointer user_data);
-  void (*set_interaction_updated_cb)(GlideSession *self, GlideSessionCallback cb, gpointer user_data);
-  void (*destroy)(GlideSession *self);
-} GlideSessionOps;
-
 struct _GlideSession {
-  const GlideSessionOps *ops;
   int refcnt;
+  GlideProcess *proc;
+  StatusService *status_service;
+  gboolean started;
+  GAsyncQueue *queue;
+  GThread *thread;
+  GMutex lock;
+  GCond cond;
+  Interaction *current;
+  GlideSessionCallback added_cb;
+  gpointer added_cb_data;
+  GlideSessionCallback updated_cb;
+  gpointer updated_cb_data;
 };
 
-static inline void glide_session_eval(GlideSession *self, Interaction *interaction) {
-  g_return_if_fail(self);
-  self->ops->eval(self, interaction);
-}
-
-static inline void glide_session_set_interaction_added_cb(GlideSession *self, GlideSessionCallback cb, gpointer user_data) {
-  g_return_if_fail(self);
-  self->ops->set_interaction_added_cb(self, cb, user_data);
-}
-
-static inline void glide_session_set_interaction_updated_cb(GlideSession *self, GlideSessionCallback cb, gpointer user_data) {
-  g_return_if_fail(self);
-  self->ops->set_interaction_updated_cb(self, cb, user_data);
-}
-
+GlideSession *glide_session_new(GlideProcess *proc, StatusService *status_service);
+void          glide_session_eval(GlideSession *self, Interaction *interaction);
+void          glide_session_set_interaction_added_cb(GlideSession *self, GlideSessionCallback cb, gpointer user_data);
+void          glide_session_set_interaction_updated_cb(GlideSession *self, GlideSessionCallback cb, gpointer user_data);
 GlideSession *glide_session_ref(GlideSession *self);
-void glide_session_unref(GlideSession *self);
+void          glide_session_unref(GlideSession *self);
 
+void          glide_session_on_message(GString *msg, gpointer user_data);
